# contracts/generated/rust/openapi.json
{
  "components": {
    "schemas": {
...
      "DeviceOut": {
        "properties": {
          "added_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "client_id": {
            "type": "string"
          },
          "connected": {
            "type": "boolean"
          },
          "is_output": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "client_id",
          "name",
          "is_output",
          "connected"
        ],
        "type": "object"
      },
      "DevicePut": {
        "properties": {
          "is_output": {
            "default": false,
            "type": "boolean"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "HTTPValidationError": {
        "description": "FastAPI-compatible validation envelope retained as a stable HTTP contract.",
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "type": "object"
          },
          "input": {},
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array"
          },
          "msg": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Music contributors"
    },
    "description": "",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "music-server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/devices/{client_id}": {
      "delete": {
        "operationId": "forget_device",
        "parameters": [
          {
            "description": "Stable client identity",
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "tags": [
          "devices"
        ]
      },
      "put": {
        "operationId": "save_device",
        "parameters": [
          {
            "description": "Stable client identity",
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevicePut"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "tags": [
          "devices"
        ]
      }
    },
...
  }
}